+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 * Python >=3.11 required
 
+### Fixed
+* Strawberry >=0.288 compatible
+
 ## [2.0.1](https://pypi.org/project/graphique/2.0.1/) - 2025-11-12
 ### Fixed
 * Strawberry >=0.285 compatible